+++ src/app/dashboard/page.js
@@ -1,0 +1,9 @@
+"use client";
+
+import React from "react";
+
+const page = () => {
+  return <div></div>;
+};
+
+export default page;
