+++ src/app/login/layout.js
@@ -1,0 +1,14 @@
+export const metadata = {
+  title: "Next.js",
+  description: "Generated by Next.js",
+};
+
+export default function LoginLayout({ children }) {
+  return (
+    <html lang="en">
+      <body className="grid h-screen place-items-center bg-gray-400">
+        {children}
+      </body>
+    </html>
+  );
+}
